# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentis Semear API",
        "description": "API do monólito Mentis Semear (autenticação JWT, workshops, check-ins, avaliações e diário).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Mesma origem (dev e produção)"
        }
    ],
    "paths": {
...
        "/api/diary-evidences": {
            "get": {
                "tags": [
                    "DiaryEvidences"
                ],
                "summary": "Lista evidências (paginado)",
                "operationId": "03925b9cbe1209e9b3d007a0bdc9552c",
                "parameters": [
                    {
                        "name": "diary_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DiaryEvidence"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "DiaryEvidences"
                ],
                "summary": "Adiciona uma evidência a um diário",
                "operationId": "36db609ba3444a174db15f3ae9130971",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "diary_id",
                                    "link"
                                ],
                                "properties": {
                                    "diary_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "link": {
                                        "type": "string",
                                        "example": "https://drive.google.com/file/abc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DiaryEvidence"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/diary-evidences/{diaryEvidence}": {
            "get": {
                "tags": [
                    "DiaryEvidences"
                ],
                "summary": "Detalha uma evidência",
                "operationId": "d810f9c5d2e4dfe998f621ba6aee4df7",
                "parameters": [
                    {
                        "name": "diaryEvidence",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Evidência",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DiaryEvidence"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "DiaryEvidences"
                ],
                "summary": "Atualiza o link de uma evidência",
                "operationId": "429d4da6a5a4a68af84eb608c90a232a",
                "parameters": [
                    {
                        "name": "diaryEvidence",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "link": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/DiaryEvidence"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "DiaryEvidences"
                ],
                "summary": "Remove uma evidência",
                "operationId": "432efaf939aa6ef0e4e934b84194dad6",
                "parameters": [
                    {
                        "name": "diaryEvidence",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DiaryEvidence": {
                "properties": {
                    "id": {
                        "description": "OpenAPI schema only — never instantiated.",
                        "type": "integer",
                        "example": 1
                    },
                    "diary_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_creator_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "link": {
                        "type": "string",
                        "example": "https://drive.google.com/file/abc"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie o token no header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}